+++ coordinator/ethereum/gas-pricing/dynamic-cap/src/main/kotlin/net/consensys/linea/ethereum/gaspricing/dynamiccap/GasPriceCapFeeHistoryFetcherImpl.kt
@@ -57,7 +57,7 @@ class GasPriceCapFeeHistoryFetcherImpl(
         "than maxBlockCount=${config.maxBlockCount}"
     }
 
-    val newestBlock = endBlockNumberInclusive.toULong().toBlockParameter()
+    val newestBlock = endBlockNumberInclusive.toBlockParameter()
     val blockCount = (endBlockNumberInclusive - startBlockNumberInclusive).inc().toInt()
 
     log.debug(
